lab 13

+++ demo/src/main/java/com/example/demo/infrastructure/AlbumsRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.infrastructure;
 import com.example.demo.domain.Albums;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlbumsRepository  extends CrudRepository <Albums, Integer> {
+public interface AlbumsRepository  extends CrudRepository <Albums, Long> {
 
 }
